pyproject.toml
[project]
name = "blog-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "fastapi>=0.121.3",
    "gradio>=6.0.1",
    "groq>=0.36.0",
    "jsonpatch>=1.33",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.1.2",
]
